% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfspav_functions.r
\name{predict.rfspav}
\alias{predict.rfspav}
\title{predict.rfspav}
\usage{
\method{predict}{rfspav}(object, new_data, cpus = detectCores() - 1, ...)
}
\arguments{
\item{object}{Model}

\item{new_data}{New data}

\item{cpus}{Number of cores, Default: detectCores() - 1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Make prediction of rfspav object
}
\details{
DETAILS
}
\seealso{

}
