% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfspav_functions.r
\name{rfspav}
\alias{rfspav}
\title{rfspav}
\usage{
rfspav(
  formula,
  data,
  coord_names,
  neighbors = 5,
  type = c("rf", "rfsp", "rfsi", "rfsig"),
  trees = 500,
  mtry = NULL,
  min.node.size = NULL,
  progress = TRUE,
  importance = "impurity",
  gower_vars = NA,
  clust_style = c("sd", "equal", "pretty", "quantile", "kmeans", "hclust"),
  num_class = 10,
  feature_select = FALSE,
  data_crs = NA,
  ...
)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data}

\item{coord_names}{Names of columns containing spatial coordinates}

\item{neighbors}{Neighbours, Default: 5}

\item{type}{Type of Random Forest Spatial Variants, Default: c("rf", "rfsp", "rfsi", "rfsig")}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{min.node.size}{Minimal node size. Default 1 for classification, 5 for regression, 3 for survival, and 10 for probability.}

\item{progress}{Progress, Default: TRUE}

\item{importance}{Type of variable importance, Default: 'impurity'}

\item{gower_vars}{Variables for gower distance calculation, Default: NA}

\item{clust_style}{Type of methods for choosing univariate class intervals of target variables (only for "rfsp" approach), Default: c("sd", "equal", "pretty", "quantile", "kmeans", "hclust")}

\item{num_class}{Number of class intervals, Default: 10}

\item{feature_select}{Should Feature Selection algorithm be applied (only for "rfsig" approach), Default: FALSE}

\item{data_crs}{Data CRS, Default: NA}

\item{...}{Additional parameters for ranger Random Forest implementation.}

\item{num.trees}{Number of trees.}

\item{cpus}{Number of cores, Default: detectCores() - 1}
}
\value{
List of the following elements: Model, Data, Formula, etc.
}
\description{
Random Forest SPatial Variants
}
\details{
DETAILS
}
\seealso{

}
