% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfspav_functions.r
\name{select_best_params}
\alias{select_best_params}
\title{select_best_params}
\usage{
select_best_params(x, metric = c("rmse", "rsq"))
}
\arguments{
\item{x}{Results of tune.rfspav function}

\item{metric}{Regression metrics, Default: c("rmse", "rsq")}
}
\value{
Best parameters
}
\description{
Select best parameters
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
