% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfspav_functions.r
\name{tune.rfspav}
\alias{tune.rfspav}
\title{Tune rfspav model (select best parameters)}
\usage{
tune.rfspav(
  formula,
  data_resample,
  coord_names,
  param_grid,
  metric = c("rmse", "rsq"),
  type = c("rf", "rfsp", "rfsi", "rfsig"),
  progress = TRUE,
  cpus = 1,
  importance = "impurity",
  gower_vars = NA,
  clust_style = c("sd", "equal", "pretty", "quantile", "kmeans", "hclust"),
  num_class = 10,
  feature_select = FALSE,
  data_crs = NA,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data_resample}{'rsample' folds}

\item{coord_names}{PARAM_DESCRIPTION}

\item{param_grid}{grid of four meta-parameters: neighbors, mtry, trees, min.node,size.}

\item{metric}{Regression metrics, Default: c("rmse", "rsq")}

\item{type}{Type of Random Forest Spatial Variants, Default: c("rf", "rfsp", "rfsi", "rfsig")}

\item{progress}{PARAM_DESCRIPTION, Default: TRUE}

\item{cpus}{Number of cores, Default: 1}

\item{importance}{Variable importance, Default: 'impurity'}

\item{gower_vars}{Variable used of gower-distance calculation, Default: NA}

\item{clust_style}{Type of methods for choosing univariate class intervals of target variables (only for "rfsp" approach), Default: c("sd", "equal", "pretty", "quantile", "kmeans", "hclust")}

\item{num_class}{Number of class intervals, Default: 10}

\item{feature_select}{Should Feature Selection algorithm be applied (only for "rfsig" approach), Default: FALSE}

\item{data_crs}{Data CRS, Default: NA}

\item{...}{Additional parameters for 'ranger' Random Forest}
}
\value{
Tuning list
}
\description{
Do cross-validation procedured for tuning meta-parameters
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
